Geschwindigkeits Item funzt wie es soll

+++ client/src/main/java/de/ostfalia/group4/client/model/Geschwindigkeitsmodifier.java
@@ -1,9 +1,9 @@
 package de.ostfalia.group4.client.model;
 
 public class Geschwindigkeitsmodifier extends Item{
-    public int geschwindigkeitsmodifier;
+    public double geschwindigkeitsmodifier;
 
-    public Geschwindigkeitsmodifier(Position position, int geschwindigkeitsmodifier) {
+    public Geschwindigkeitsmodifier(Position position, double geschwindigkeitsmodifier) {
         this.position=position;
         this.geschwindigkeitsmodifier=geschwindigkeitsmodifier;
     }
